Module pallet_fund_admin::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

Functions

approve_drawdown

Approve a drawdown

approve_revenue

Approve a revenue

bank_confirming_documents

The following extrinsic is used to upload the bank confirming documents for a given drawdown.

expenditures_and_job_eligibles

This extrinsic is used to create, update or delete expenditures & job eligibles.

inflation_rate

Modifies the inflation rate of a project.

initial_setup

Initialize the pallet by setting the permissions for each role & the global scope

kill_storage

Kill all the stored data.

projects_assign_user

Assigns a user to a project.

projects_create_project

Registers a new project.

projects_delete_project

Deletes a project.

projects_edit_project

Edits a project.

recovery_drawdown

Execute a recovery drawdown on a project. This function can only be called by an admin.

recovery_revenue

Execute a recovery revenue on a project. This function can only be called by an admin.

reject_drawdown

Reject a drawdown

reject_revenue

Reject a revenue

reset_drawdown

The following extrinsic is used to cancel a drawdown submission.

set_new_admin_permissions

submit_drawdown

Submit a drawdown This extrinsic is used to create, update or delete transactions. It also allows that an array of transactions to be saved as a draft or as submitted.

submit_revenue

This extrinsic is used to create, update or delete revenue transactions. It also allows that an array of revenue transactions to be saved as a draft or as submitted.

sudo_add_administrator

Adds an administrator account to the site

sudo_remove_administrator

Removes an administrator account from the site

up_bulkupload

Bulk upload drawdowns.

users

This extrinsic is used to create, update, or delete a user account

users_edit_user

Edits an user account