Function pallet_fund_admin::pallet::dispatchables::expenditures_and_job_eligibles

pub fn expenditures_and_job_eligibles<T: Config>(
    project_id: [u8; 32],
    expenditures: Option<BoundedVec<(Option<BoundedVec<u8, ConstU32<100>>>, Option<ExpenditureType>, Option<u64>, Option<BoundedVec<u8, ConstU32<400>>>, Option<u32>, CUDAction, Option<[u8; 32]>), <T as Config>::MaxRegistrationsAtTime>>,
    job_eligibles: Option<BoundedVec<(Option<BoundedVec<u8, ConstU32<100>>>, Option<u64>, Option<BoundedVec<u8, ConstU32<400>>>, Option<u32>, CUDAction, Option<[u8; 32]>), <T as Config>::MaxRegistrationsAtTime>>
)

This extrinsic is used to create, update or delete expenditures & job eligibles.

Parameters:

origin: The administrator account

project_id: The selected project id where the expenditures will be created/updated/deleted

expenditures: The expenditures to be created/updated/deleted. This is a vector of tuples where each entry is composed by:

0: The name of the expenditure

1: The expenditure type

2: The amount of the expenditure

3: The naics code of the expenditure

4: The jobs multiplier of the expenditure

5: The expenditure action to be performed. (Create, Update or Delete)

6: The expenditure id. This is only used when updating or deleting an expenditure.

job_eligibles: The job eligibles to be created/updated/deleted. This is a vector of tuples where each entry is composed by:

0: The job eligible name

1: The amount of the job eligible

2: The NAICS code of the job eligible

3: The jobs multiplier of the job eligible

4: The job eligible action to be performed. (Create, Update or Delete)

5: The job eligible id. This is only used when updating or deleting a job eligible.

Considerations:

Naics code and jobs multiplier are always optional.

This function can only be called by an administrator account

This extrinsic allows multiple expenditures to be created/updated/deleted at the same time.

The project id is required because it is the only way to identify the project

Expenditure parameters are optional because depends on the action to be performed:

Create: Name, Type & Amount are required. Nacis code & Jobs multiplier are optional.

Update: Except for the expenditure id & action, all parameters are optional.

Delete: Only the expenditure id & action is required.

Multiple actions can be performed at the same time. For example, you can create a new expenditure and update another one at the same time.

Do not perform multiple actions over the same expenditure in the same call, it could result in an unexpected behavior.

NOTE: This function is an automatically generated, doc only, uncallable stub. designed to document the expenditures_and_job_eligibles variant of the pallet Call enum. You should not attempt to call this function directly.