Function pallet_fund_admin::pallet::dispatchables::projects_create_project

pub fn projects_create_project<T: Config>(
    title: BoundedVec<u8, ConstU32<100>>,
    description: BoundedVec<u8, ConstU32<400>>,
    image: Option<BoundedVec<u8, ConstU32<100>>>,
    address: BoundedVec<u8, ConstU32<100>>,
    banks: Option<BoundedVec<(BoundedVec<u8, ConstU32<100>>, BoundedVec<u8, ConstU32<100>>), T::MaxBanksPerProject>>,
    creation_date: u64,
    completion_date: u64,
    expenditures: BoundedVec<(Option<BoundedVec<u8, ConstU32<100>>>, Option<ExpenditureType>, Option<u64>, Option<BoundedVec<u8, ConstU32<400>>>, Option<u32>, CUDAction, Option<[u8; 32]>), <T as Config>::MaxRegistrationsAtTime>,
    job_eligibles: Option<BoundedVec<(Option<BoundedVec<u8, ConstU32<100>>>, Option<u64>, Option<BoundedVec<u8, ConstU32<400>>>, Option<u32>, CUDAction, Option<[u8; 32]>), <T as Config>::MaxRegistrationsAtTime>>,
    users: Option<BoundedVec<(<T as Config>::AccountId, ProxyRole, AssignAction), <T as Config>::MaxRegistrationsAtTime>>,
    private_group_id: BoundedVec<u8, ConstU32<400>>
)

Registers a new project.

Parameters:

origin: The administrator account

title: The title of the project

description: The description of the project

image: The image of the project (CID)

address: The address of the project

creation_date: The creation date of the project

completion_date: The completion date of the project

expenditures: The expenditures of the project. It is an array of tuples where each entry is a tuple of the following:

0: The expenditure name

1: The expenditure type

2: The expenditure amount

3: The expenditure NAICS code

4: The expenditure jobs multiplier

5: The CUD action to be performed on the expenditure. CUD action is ALWAYS required.

6: The expenditure id. It is optional because it is only required when updating or deleting

job_eligibles: The job eligibles to be created/updated/deleted. This is a vector of tuples where each entry is composed by:

0: The job eligible name

1: The amount of the job eligible

2: The NAICS code of the job eligible

3: The jobs multiplier of the job eligible

4: The job eligible action to be performed. (Create, Update or Delete)

5: The job eligible id. This is only used when updating or deleting a job eligible.

users: The users who will be assigned to the project. It is an array of tuples where each entry is a tuple of the following:

0: The user account

1: The user role

2: The AssignAction to be performed on the user.

Considerations:

This function can only be called by an administrator account

For users assignation, the user account must be registered. If the user is not registered, the function will return an error. ALL parameters are required.

For expenditures, apart from the expenditure id, naics code & jopbs multiplier, ALL parameters are required because for this flow, the expenditures are always created. The naics code & the jobs multiplier can be added later by the administrator.

Creating a project will automatically create a scope for the project.

Note:

WARNING: If users are provided, the function will assign the users to the project, granting them permissions in the rbac pallet.

NOTE: This function is an automatically generated, doc only, uncallable stub. designed to document the projects_create_project variant of the pallet Call enum. You should not attempt to call this function directly.