Function pallet_fund_admin::pallet::dispatchables::submit_revenue

pub fn submit_revenue<T: Config>(
    project_id: [u8; 32],
    revenue_id: [u8; 32],
    revenue_transactions: Option<BoundedVec<(Option<[u8; 32]>, Option<u64>, Option<BoundedVec<(BoundedVec<u8, ConstU32<100>>, BoundedVec<u8, ConstU32<100>>), <T as Config>::MaxDocuments>>, CUDAction, Option<[u8; 32]>), <T as Config>::MaxRegistrationsAtTime>>,
    submit: bool
)

This extrinsic is used to create, update or delete revenue transactions. It also allows that an array of revenue transactions to be saved as a draft or as submitted.

Parameters:

*/ - origin: The user account who is creating the revenue transactions

project_id: The selected project id where the revenue transactions will be created

revenue_id: The selected revenue id where the revenue transactions will be created

revenue_transactions: The revenue transactions to be created/updated/deleted. This entry is a vector of tuples where each entry is composed by:

0: The job eligible id where the revenue transaction will be created

1: The amount of the revenue transaction

2: Documents associated to the revenue transaction

3: The action to be performed on the revenue transaction (Create, Update or Delete)

4: The revenue transaction id. This is required only if the action is being updated or deleted.

submit: If true, the array of revenue transactions will be submitted to the administrator. If false, the array of revenue transactions will be saved as a draft.

Considerations:

This function is only callable by a builder role account

This extrinsic allows multiple revenue transactions to be created/updated/deleted at the same time.

The project id and revenue id are required for the reports.

revenue_transactions parameters are optional because depends on the action to be performed:

Create: Job eligible id, Amount, Documents & action are required.

Update: Except for the revenue transaction id & action, all other parameters are optional.

Delete: Only the revenue transaction id & action are required.

Multiple actions can be performed at the same time, but each must be performed on a different transaction. For example, you can create a new transaction and update another one at the same time.

Do not perform multiple actions over the same transaction in the same call, it could result in an unexpected behavior.

If a revenue is submitted, all transactions must be submitted too. If the revenue do not contain any transaction, it will return an error.

After a revenue is submitted, it can not be updated or deleted.

After a revenue is rejected, builders will use again this extrinsic to update the transactions associated to a given revenue.

NOTE: This function is an automatically generated, doc only, uncallable stub. designed to document the submit_revenue variant of the pallet Call enum. You should not attempt to call this function directly.