Function pallet_fund_admin::pallet::dispatchables::users

pub fn users<T: Config>(
    users: BoundedVec<(<T as Config>::AccountId, Option<BoundedVec<u8, ConstU32<100>>>, Option<ProxyRole>, CUDAction), <T as Config>::MaxRegistrationsAtTime>
)

This extrinsic is used to create, update, or delete a user account

Parameters:

origin: The administrator account

user: The target user account to be registered, updated, or deleted. It is an array of user accounts where each entry it should be a tuple of the following:

0: The user account

1: The user name

2: The user role

3: The CUD operation to be performed on the user account. CUD action is ALWAYS required.

Considerations:

Users parameters are optional because depends on the CUD action as follows:

Create: The user account, user name, user role & CUD action are required

Update: The user account & CUD action are required. The user name & user role are optionals.

Delete: The user account & CUD action are required.

This function can only be called by an administrator account

Multiple users can be registered, updated, or deleted at the same time, but the user account must be unique. Multiple actions over the same user account in the same call, it could result in an unexpected behavior.

If the user is already registered, the function will return an error: UserAlreadyRegistered

If the user is not registered, the function will return an error: UserNotFound

Note:

WARNING: It is possible to register, update, or delete administrators accounts using this extrinsic, but administrators can not delete themselves.

WARNING: This function only registers, updates, or deletes users from the site.

WARNING: The only way to grant or remove permissions of a user account is assigning or unassigning a user from a selected project.

NOTE: This function is an automatically generated, doc only, uncallable stub. designed to document the users variant of the pallet Call enum. You should not attempt to call this function directly.