Enum pallet_fund_admin::pallet::Error

pub enum Error<T> {

    EmptyFieldName,
    EmptyFieldDescription,
    FieldNameTooLong,
    EmptyUsers,
    EmptyFieldCID,
    EmptyFieldBanks,
    PrivateGroupIdEmpty,
    EmptyUsersAssignation,
    EmptyProjectAddress,
    NoGlobalScopeValueWasFound,
    ProjectIdAlreadyInUse,
    TimestampError,
    CompletionDateMustBeLater,
    UserAlreadyRegistered,
    ProjectNotFound,
    ProjectIsAlreadyCompleted,
    ProjectHasNoDrawdowns,
    ProjectHasNoExpenditures,
    ProjectHasNoUsers,
    ProjectHasNoJobEligibles,
    ProjectHasNoRevenues,
    CannotDeleteCompletedProject,
    UserNotRegistered,
    UserAlreadyAssignedToProject,
    MaxUsersPerProjectReached,
    MaxProjectsPerUserReached,
    UserNotAssignedToProject,
    CannotRegisterAdminRole,
    MaxBuildersPerProjectReached,
    MaxInvestorsPerProjectReached,
    MaxIssuersPerProjectReached,
    MaxRegionalCenterPerProjectReached,
    CannotRemoveAdminRole,
    CannotAddAdminRole,
    UserCannotHaveMoreThanOneRole,
    ExpenditureNotFound,
    ExpenditureNotFoundForSelectedProjectId,
    ExpenditureAlreadyExists,
    ExpenditureHasNonZeroTransactions,
    MaxExpendituresPerProjectReached,
    EmptyExpenditureName,
    ExpenditureDoesNotBelongToProject,
    DrawdownNotFound,
    InvalidAmount,
    DocumentsEmpty,
    TransactionNotFound,
    TransactionNotFoundForSelectedDrawdownId,
    TransactionAlreadyExists,
    TransactionInUse,
    MaxTransactionsPerDrawdownReached,
    DrawdownAlreadyExists,
    MaxDrawdownsPerProjectReached,
    MaxStatusChangesPerDrawdownReached,
    MaxRecoveryChangesReached,
    CannotEditDrawdown,
    CannotPerformActionOnSubmittedTransaction,
    CannotPerformActionOnApprovedTransaction,
    CannotPerformActionOnConfirmedTransaction,
    CannotPerformActionOnSubmittedDrawdown,
    CannotPerformActionOnApprovedDrawdown,
    CannotPerformActionOnConfirmedDrawdown,
    TransactionIsAlreadyCompleted,
    UserDoesNotHaveRole,
    EmptyTransactions,
    TransactionsRequired,
    TransactionIdRequired,
    DrawdownHasNoTransactions,
    CannotSubmitTransaction,
    DrawdownIsNotInSubmittedStatus,
    TransactionIsNotInSubmittedStatus,
    EmptyExpenditures,
    ExpenditureNameRequired,
    ExpenditureTypeRequired,
    ExpenditureAmountRequired,
    ExpenditureIdRequired,
    UserNameRequired,
    UserRoleRequired,
    UserImageRequired,
    UserEmailRequired,
    AmountRequired,
    UserHasAssignedProjects,
    UserHasNoProjects,
    NoTransactionsToSubmit,
    BulkUploadDescriptionRequired,
    BulkUploadDocumentsRequired,
    AdministratorsCannotDeleteThemselves,
    NoFeedbackProvidedForBulkUpload,
    EmptyBulkUploadFeedback,
    EB5MissingFeedback,
    EmptyEb5Feedback,
    ProjectsInflationRateEmpty,
    InflationRateRequired,
    InflationRateAlreadySet,
    InflationRateNotSet,
    DrawdownTypeNotSupportedForBulkUpload,
    UserHasAssignedProjectsCannotUpdateRole,
    UserHasAssignedProjectsCannotDelete,
    DrawdownStatusNotSupportedForBulkUpload,
    DrawdownIsNotInDraftOrRejectedStatus,
    UserIsNotAnInvestor,
    MaxProjectsPerInvestorReached,
    JobEligiblesEmpty,
    JobEligiblesNameRequired,
    JobEligibleIdAlreadyExists,
    MaxJobEligiblesPerProjectReached,
    JobEligibleNotFound,
    JobEligibleDoesNotBelongToProject,
    JobEligibleNameRequired,
    JobEligibleAmountRequired,
    JobEligibleIdRequired,
    JobEligibleNotFoundForSelectedProjectId,
    JobEligibleHasNonZeroTransactions,
    RevenueNotFound,
    RevenueTransactionsEmpty,
    RevenueTransactionsRequired,
    RevenueTransactionNotSubmitted,
    CannotEditRevenue,
    RevenueTransactionIdAlreadyExists,
    MaxTransactionsPerRevenueReached,
    RevenueTransactionNotFound,
    RevenueTransactionNotFoundForSelectedRevenueId,
    CannotEditRevenueTransaction,
    MaxStatusChangesPerRevenueReached,
    CannotPerformActionOnSubmittedRevenue,
    CannotPerformActionOnApprovedRevenue,
    CannotPerformActionOnApprovedRevenueTransaction,
    CannotPerformActionOnSubmittedRevenueTransaction,
    RevenueAmountRequired,
    RevenueTransactionIdRequired,
    RevenueIdAlreadyExists,
    MaxRevenuesPerProjectReached,
    RevenueHasNoTransactions,
    RevenueIsNotInSubmittedStatus,
    RevenueTransactionIsNotInSubmittedStatus,
    RevenueTransactionsFeedbackEmpty,
    RevenueNotSubmitted,
    RevenueDoesNotBelongToProject,
    DrawdowMustBeInApprovedStatus,
    DrawdowMustBeInConfirmedStatus,
    DrawdownNotSubmitted,
    DrawdownHasAlreadyBankConfirmingDocuments,
    DrawdownHasNoBankConfirmingDocuments,
    DrawdownDoesNotBelongToProject,
    BankConfirmingDocumentsNotProvided,
    BankConfirmingDocumentsEmpty,
    OnlyEB5DrawdownsCanUploadBankDocuments,
    MaxRegistrationsAtATimeReached,
    AdminHasNoFreeBalance,
    InsufficientFundsToTransfer,
    // some variants omitted

}

Custom dispatch errors of this pallet.

Variants

EmptyFieldName

FieldName is empty

EmptyFieldDescription

FieldDescription is empty

FieldNameTooLong

FieldName is too long

EmptyUsers

Array of users is empty

EmptyFieldCID

CID is empty

EmptyFieldBanks

Array of banks is empty

PrivateGroupIdEmpty

The private group id is empty

EmptyUsersAssignation

Array of users to be assigned to a project is empty

EmptyProjectAddress

Field address project is empty

NoGlobalScopeValueWasFound

No value was found for the global scope

ProjectIdAlreadyInUse

Project ID is already in use

TimestampError

Timestamp was not genereated correctly

CompletionDateMustBeLater

Completion date must be later than creation date

UserAlreadyRegistered

User is already registered in the site

ProjectNotFound

Project was not found

ProjectIsAlreadyCompleted

Project is not active anymore

ProjectHasNoDrawdowns

Project has no drawdowns

ProjectHasNoExpenditures

Project has no expenditures

ProjectHasNoUsers

Project has no users

ProjectHasNoJobEligibles

Project has no job eligibles

ProjectHasNoRevenues

Project has no revenues

CannotDeleteCompletedProject

Can not delete a completed project

UserNotRegistered

User is not registered

UserAlreadyAssignedToProject

User has been already added to the project

MaxUsersPerProjectReached

Max number of users per project reached

MaxProjectsPerUserReached

Max number of projects per user reached

UserNotAssignedToProject

User is not assigned to the project

CannotRegisterAdminRole

Can not register administrator role

MaxBuildersPerProjectReached

Max number of builders per project reached

MaxInvestorsPerProjectReached

Max number of investors per project reached

MaxIssuersPerProjectReached

Max number of issuers per project reached

MaxRegionalCenterPerProjectReached

Max number of regional centers per project reached

CannotRemoveAdminRole

Can not remove administrator role

CannotAddAdminRole

Can not add admin role at user project assignment

UserCannotHaveMoreThanOneRole

User can not have more than one role at the same time

ExpenditureNotFound

Expenditure not found

ExpenditureNotFoundForSelectedProjectId

Expenditure not found for the selected project_id

ExpenditureAlreadyExists

Expenditure already exist

ExpenditureHasNonZeroTransactions

Expenditure is already in a transaction

MaxExpendituresPerProjectReached

Max number of expenditures per project reached

EmptyExpenditureName

Field name can not be empty

ExpenditureDoesNotBelongToProject

Expenditure does not belong to the project

DrawdownNotFound

Drawdown id is not found

InvalidAmount

Invalid amount

DocumentsEmpty

Documents field is empty

TransactionNotFound

Transaction id is not found

TransactionNotFoundForSelectedDrawdownId

Transaction was not found for the selected Drawdown_id

TransactionAlreadyExists

Transaction already exist

TransactionInUse

Transaction is already in a drawdown

MaxTransactionsPerDrawdownReached

Max number of transactions per drawdown reached

DrawdownAlreadyExists

Drawdown already exist

MaxDrawdownsPerProjectReached

Max number of drawdowns per project reached

MaxStatusChangesPerDrawdownReached

Max number of status changes per drawdown reached

MaxRecoveryChangesReached

Max number of recovery chnages per drawdown reached

CannotEditDrawdown

Can not modify a completed drawdown

CannotPerformActionOnSubmittedTransaction

Can not perform any action on a submitted transaction

CannotPerformActionOnApprovedTransaction

Can not perform any action on a approved transaction

CannotPerformActionOnConfirmedTransaction

Can not perform any action on a confirmed transaction

CannotPerformActionOnSubmittedDrawdown

Can not perform any action on a submitted drawdown

CannotPerformActionOnApprovedDrawdown

Can not perform any action on a approved drawdown

CannotPerformActionOnConfirmedDrawdown

Can not perform any action on a confirmed drawdown

TransactionIsAlreadyCompleted

Transaction is already completed

UserDoesNotHaveRole

User does not have the specified role

EmptyTransactions

Transactions vector is empty

TransactionsRequired

Transactions are required for the current workflow

TransactionIdRequired

Transaction ID was not found in do_execute_transaction

DrawdownHasNoTransactions

Drawdown can not be submitted if does not has any transactions

CannotSubmitTransaction

Cannot submit transaction

DrawdownIsNotInSubmittedStatus

Drawdown can not be approved if is not in submitted status

TransactionIsNotInSubmittedStatus

Transactions is not in submitted status

EmptyExpenditures

Array of expenditures is empty

ExpenditureNameRequired

Expenditure name is required

ExpenditureTypeRequired

Expenditure type is required

ExpenditureAmountRequired

Expenditure amount is required

ExpenditureIdRequired

Expenditure id is required

UserNameRequired

User name is required

UserRoleRequired

User role is required

UserImageRequired

User image is required

UserEmailRequired

User email is required

AmountRequired

Amount is required

UserHasAssignedProjects

Can not delete a user if the user is assigned to a project

UserHasNoProjects

User has no projects assigned

NoTransactionsToSubmit

Can not send a drawdown to submitted status if it has no transactions

BulkUploadDescriptionRequired

Bulk upload description is required

BulkUploadDocumentsRequired

Bulk upload documents are required

AdministratorsCannotDeleteThemselves

Administrator can not delete themselves

NoFeedbackProvidedForBulkUpload

No feedback was provided for bulk upload

EmptyBulkUploadFeedback

Bulkupload feedback is empty

EB5MissingFeedback

NO feedback for EN5 drawdown was provided

EmptyEb5Feedback

EB5 feedback is empty

ProjectsInflationRateEmpty

Inflation rate extrinsic is missing an array of project ids

InflationRateRequired

Inflation rate was not provided

InflationRateAlreadySet

Inflation rate has been already set for the selected project

InflationRateNotSet

Inflation rate was not set for the selected project

DrawdownTypeNotSupportedForBulkUpload

Bulkupload drawdowns are only allowed for Construction Loan & Developer Equity

UserHasAssignedProjectsCannotUpdateRole

Cannot edit user role if the user is assigned to a project

UserHasAssignedProjectsCannotDelete

Cannot delete user if the user is assigned to a project

DrawdownStatusNotSupportedForBulkUpload

Cannot send a bulkupload drawdown if the drawdown status isn’t in draft or rejected

DrawdownIsNotInDraftOrRejectedStatus

Cannot submit a drawdown if the drawdown status isn’t in draft or rejected

UserIsNotAnInvestor

Only investors can update/edit their documents

MaxProjectsPerInvestorReached

Max number of projects per investor has been reached

JobEligiblesEmpty

Jobs eligibles array is empty

JobEligiblesNameRequired

JOb eligible name is empty

JobEligibleIdAlreadyExists

Job eligible id already exists

MaxJobEligiblesPerProjectReached

Max number of job eligibles per project reached

JobEligibleNotFound

Job eligible id not found

JobEligibleDoesNotBelongToProject

Jopb eligible does not belong to the project

JobEligibleNameRequired

Job eligible name is required

JobEligibleAmountRequired

Job eligible amount is required

JobEligibleIdRequired

Job eligible id is required

JobEligibleNotFoundForSelectedProjectId

Job eligible not found for the given project id

JobEligibleHasNonZeroTransactions

Job eligible has non zero transactions

RevenueNotFound

Revenue id was not found

RevenueTransactionsEmpty

Transactions revenue array is empty

RevenueTransactionsRequired

An array of revenue transactions is required

RevenueTransactionNotSubmitted

Revenue transaction is not in submitted status

CannotEditRevenue

Revenue can not be edited

RevenueTransactionIdAlreadyExists

Revenue transaction id already exists

MaxTransactionsPerRevenueReached

Max number of transactions per revenue reached

RevenueTransactionNotFound

Revenue transaction id not found

RevenueTransactionNotFoundForSelectedRevenueId

Revenue transaction was not found for the selected revenue_id

CannotEditRevenueTransaction

Revenue transaction can not be edited

MaxStatusChangesPerRevenueReached

Max number of status changes per revenue reached

CannotPerformActionOnSubmittedRevenue

Can not perform any action on a submitted revenue

CannotPerformActionOnApprovedRevenue

Can not perform any action on a approved revenue

CannotPerformActionOnApprovedRevenueTransaction

Can not perform any action on a submitted revenue transaction

CannotPerformActionOnSubmittedRevenueTransaction

Can not perform any action on a approved revenue transaction

RevenueAmountRequired

Revenue amoun is required

RevenueTransactionIdRequired

Revenue transaction id is required

RevenueIdAlreadyExists

Revenue Id already exists

MaxRevenuesPerProjectReached

Maximun number of revenues per project reached

RevenueHasNoTransactions

Can not send a revenue to submitted status if it has no transactions

RevenueIsNotInSubmittedStatus

Revenue is not in submitted status

RevenueTransactionIsNotInSubmittedStatus

Revenue transaction is not in submitted status

RevenueTransactionsFeedbackEmpty

Revenue transactions feedback is empty

RevenueNotSubmitted

The revenue is not in submitted status

RevenueDoesNotBelongToProject

The revenue id does not belong to the project

DrawdowMustBeInApprovedStatus

Can not upload bank confirming documents if the drawdown is not in Approved status

DrawdowMustBeInConfirmedStatus

Drawdown is not in Confirmed status

DrawdownNotSubmitted

Drawdown is not in Submitted status

DrawdownHasAlreadyBankConfirmingDocuments

Can not insert (CUDAction: Create) bank confmirng documents if the drawdown has already bank confirming documents

DrawdownHasNoBankConfirmingDocuments

Drawdown has no bank confirming documents (CUDAction: Update or Delete)

DrawdownDoesNotBelongToProject

Drawdown id does not belong to the selected project

BankConfirmingDocumentsNotProvided

Bank confirming documents are required

BankConfirmingDocumentsEmpty

Banck confirming documents array is empty

OnlyEB5DrawdownsCanUploadBankDocuments

Only eb5 drawdowns are allowed to upload bank documentation

MaxRegistrationsAtATimeReached

Maximun number of registrations at a time reached

AdminHasNoFreeBalance

Administrator account has insuficiente balance to register a new user

InsufficientFundsToTransfer

Administrator account has insuficiente balance to register a new user